requirements.txt
# Windows Search Tool - Dependencies
# Python 3.11+

# GUI Framework
PyQt6==6.6.1
PyQt6-Qt6==6.6.1

# Office Document Parsing
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.16.3

# OCR
pytesseract==0.3.10
Pillow==10.1.0

# Chinese Word Segmentation
jieba==0.42.1

# Encoding Detection
chardet==5.2.0

# HTTP Client for AI API
httpx==0.25.2

# Database
# SQLite3 is included in Python standard library

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.12.1
pylint==3.0.3
mypy==1.7.1

# Type Stubs
types-requests==2.31.0.20231231
